Properly handle UltraStar's ~ metacharacter
The tilde is UltraStar's non-breaking space character, and is used to
specify different pitches for parts of the same syllable. Since we don't
have any use for it and don't want it to show up in BlitzLoop files, we
mask it and let the duration of these pseudo-syllables get carried over
to the next real syllable.